+++ evtloop.h
@@ -53,6 +53,8 @@ typedef struct {
 	uint64_t el_radar42784847;
 	uint64_t el_radar42946744_fatal;
 	uint64_t el_radar42946744;
+	uint64_t el_radar43151662_fatal;
+	uint64_t el_radar43151662;
 	uint64_t el_missingtoken;
 	uint64_t el_ooms;
 	aupipe_stat_t ap;